Fix potential crash due to null pointer...
If a tile map is already open, and the tile size form is opened but the user clicks the cancel, the tile map in the editor is set to nullptr. Clicking on the tile map to paint a tile tries to access a tile map through a null pointer which crashes the program.